refactor goal manager into  context
